Added proper country reading from response from name

+++ F1Info/MockDataGenerator/src/main/java/mock_data_generator/tables/02_f1_info_tables.sql
@@ -37,8 +37,8 @@ create table circuits(
   name varchar(255) not null unique,
   location_name varchar(255),
   country_code varchar(255),
-  latitude numeric,
-  longitude numeric,
+  latitude float,
+  longitude float,
   wikipedia_page varchar(255) not null,
 
   constraint circuits_pk primary key (id, circuit_identifier),
